% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract.values.f}
\alias{extract.values.f}
\title{Automated Data Extraction Functions for Saudi Aramco Prospectus
Tidy PDF Scraping and Fuzzy Joining Helper}
\usage{
\method{extract}{values.f}(pdf.file, page, names)
}
\arguments{
\item{names}{}
}
\value{

}
\description{
Automated Data Extraction Functions for Saudi Aramco Prospectus
Tidy PDF Scraping and Fuzzy Joining Helper
}
